backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
openai
qdrant-client
supabase
python-dotenv
PyPDF2
